[svn-r4327] Purpose:     More code cleanups Description:     Wrap up the code cleanups for changing the dataset transfer property lists     over to using the generic property list code. Platforms tested:     IRIX64 6.5 (modi4)

+++ hl/tools/gif2h5/decompress.c
@@ -162,7 +162,7 @@ Decompress(GIFIMAGEDESC *GifImageDesc, GIFHEAD *GifHead)
     OutCount = 0;
     BitOffset = 0;
     
-    DataMask = (WORD)((1L << ((GifHead->PackedField & 0x07) +1)) -1);
+    DataMask = (1L << ((GifHead->PackedField & 0x07) +1)) -1;
     Raster = GifImageDesc->GIFImage;
 
     /* Check for image seperator */
